+++ contracts/test/AllStandardOpsTest.sol
@@ -70,8 +70,8 @@ contract AllStandardOpsTest is RainVM {
         return _state;
     }
 
-    function fnPtrs() public pure override returns (uint[] memory) {
-        return AllStandardOps.fnPtrs(new uint[](0));
+    function fnPtrs() public pure override returns (uint256[] memory) {
+        return AllStandardOps.fnPtrs(new uint256[](0));
     }
 
     /// Runs `eval` and stores full state.
